remove useless content component, clean up design slightly
